+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Update `paragraphModifier` props to `MarkdownPage` component.
+
 ## [1.18.0] - 2019-02-27
 
 Features: